+++ base/src/main/kotlin/com/freeletics/gradle/plugin/FreeleticsBaseExtension.kt
@@ -2,6 +2,7 @@ package com.freeletics.gradle.plugin
 
 import com.freeletics.gradle.setup.configureDagger
 import com.freeletics.gradle.setup.configureMoshi
+import com.freeletics.gradle.setup.setupCompose
 import com.freeletics.gradle.util.kotlin
 import org.gradle.api.Project
 import org.gradle.api.plugins.ExtensionAware
@@ -20,6 +21,10 @@ public abstract class FreeleticsBaseExtension(private val project: Project) : Ex
                 optIn.addAll(*classes)
             }
         }
+    }
+
+    public fun useCompose() {
+        project.setupCompose()
     }
 
     @JvmOverloads
